backend/requirements.txt
flask>=2.0.0
flask-sqlalchemy>=3.0.0
flask-jwt-extended>=4.0.0
flask-cors>=4.0.0
python-dotenv>=0.19.0
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0
pymysql>=1.0.0
cryptography>=3.4.0
